separate enricher from arg_scan for better oragnization
