+++ pkg/links/azure/arg_template.go
@@ -1,14 +1,10 @@
 package azure
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
-	"net/http"
-	"net/url"
 	"path/filepath"
 	"strings"
-	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/resourcegraph/armresourcegraph"
 	"github.com/praetorian-inc/janus-framework/pkg/chain"
@@ -27,203 +23,6 @@ import (
 type ARGTemplateQueryInput struct {
 	Template     *templates.ARGQueryTemplate
 	Subscription string
-}
-
-// Command represents the input and output of a command that requires manual triage
-type Command struct {
-	Command                   string `json:"command"`
-	Description               string `json:"description"`
-	ExpectedOutputDescription string `json:"expected_output_description"`
-	ActualOutput              string `json:"actual_output"`
-	ExitCode                  int    `json:"exit_code"`
-	Error                     string `json:"error,omitempty"`
-}
-
-// ResourceEnricher interface for extensible resource enrichment
-type ResourceEnricher interface {
-	CanEnrich(templateID string) bool
-	Enrich(ctx context.Context, resource *model.AzureResource) []Command
-}
-
-// StorageAccountEnricher implements enrichment for storage accounts
-type StorageAccountEnricher struct{}
-
-func (s *StorageAccountEnricher) CanEnrich(templateID string) bool {
-	return templateID == "storage_accounts_public_access"
-}
-
-func (s *StorageAccountEnricher) Enrich(ctx context.Context, resource *model.AzureResource) []Command {
-	commands := []Command{}
-
-	// Extract storage account name from resource
-	storageAccountName := resource.Name
-	if storageAccountName == "" {
-		commands = append(commands, Command{
-			Command:      "",
-			Description:  "No storage account name found",
-			ActualOutput: "Error: storage account name is empty",
-		})
-		return commands
-	}
-
-	// Sanitize the storage account name for URL encoding
-	storageAccountNameForURL := url.QueryEscape(strings.TrimSpace(storageAccountName))
-
-	// Create HTTP client with timeout
-	client := &http.Client{
-		Timeout: 10 * time.Second,
-	}
-
-	// Test anonymous access via HTTP request
-	testURL := fmt.Sprintf("https://%s.blob.core.windows.net/?comp=list", storageAccountNameForURL)
-
-	resp, err := client.Get(testURL)
-
-	command := fmt.Sprintf("curl -w \"\\n===== Status Code =====\\n%%{http_code}\\n\" \"%s\" --max-time 10", testURL)
-	curlCommand := Command{
-		Command:                   command,
-		Description:               "Test anonymous access to storage account container listing",
-		ExpectedOutputDescription: "anonymous access enabled 404 | anonymous access disabled = 401/403 | public access disabled = 409",
-	}
-
-	if err != nil {
-		curlCommand.Error = err.Error()
-		curlCommand.ActualOutput = fmt.Sprintf("Request failed: %s", err.Error())
-	} else {
-		defer resp.Body.Close()
-		curlCommand.ActualOutput = fmt.Sprintf("HTTP %d", resp.StatusCode)
-		curlCommand.ExitCode = resp.StatusCode
-	}
-
-	commands = append(commands, curlCommand)
-	return commands
-}
-
-// EnrichmentRegistry holds all available enrichers
-type EnrichmentRegistry struct {
-	enrichers []ResourceEnricher
-}
-
-func NewEnrichmentRegistry() *EnrichmentRegistry {
-	return &EnrichmentRegistry{
-		enrichers: []ResourceEnricher{
-			&StorageAccountEnricher{},
-			&VirtualMachineEnricher{}, // Example of additional enricher
-			// Add more enrichers here as needed
-		},
-	}
-}
-
-// VirtualMachineEnricher implements enrichment for virtual machines
-type VirtualMachineEnricher struct{}
-
-func (v *VirtualMachineEnricher) CanEnrich(templateID string) bool {
-	return templateID == "virtual_machines_public" || templateID == "virtual_machines_all"
-}
-
-func (v *VirtualMachineEnricher) Enrich(ctx context.Context, resource *model.AzureResource) []Command {
-	commands := []Command{}
-
-	// Extract VM name and location
-	vmName := resource.Name
-	location := resource.Region
-
-	if vmName == "" {
-		commands = append(commands, Command{
-			Command:      "",
-			Description:  "No VM name found",
-			ActualOutput: "Error: VM name is empty",
-		})
-		return commands
-	}
-
-	// Add nmap command for network scanning
-	nmapCommand := Command{
-		Command:                   fmt.Sprintf("nmap -sS -O -A %s", vmName),
-		Description:               "Network scan of the virtual machine",
-		ExpectedOutputDescription: "Open ports and services running on the VM",
-		ActualOutput:              "Manual execution required",
-	}
-	commands = append(commands, nmapCommand)
-
-	// Add SSH connection test if applicable
-	if location != "" {
-		sshCommand := Command{
-			Command:                   fmt.Sprintf("ssh -o ConnectTimeout=10 azureuser@%s.%s.cloudapp.azure.com", vmName, location),
-			Description:               "Test SSH connectivity to the VM",
-			ExpectedOutputDescription: "Connection success/failure, authentication method",
-			ActualOutput:              "Manual execution required",
-		}
-		commands = append(commands, sshCommand)
-	}
-
-	return commands
-}
-
-func (r *EnrichmentRegistry) EnrichResource(ctx context.Context, templateID string, resource *model.AzureResource) []Command {
-	var allCommands []Command
-
-	for _, enricher := range r.enrichers {
-		if enricher.CanEnrich(templateID) {
-			commands := enricher.Enrich(ctx, resource)
-			allCommands = append(allCommands, commands...)
-		}
-	}
-
-	return allCommands
-}
-
-// ARGEnrichmentLink enriches Azure resources with additional security testing commands
-type ARGEnrichmentLink struct {
-	*chain.Base
-	registry *EnrichmentRegistry
-}
-
-func NewARGEnrichmentLink(configs ...cfg.Config) chain.Link {
-	l := &ARGEnrichmentLink{
-		registry: NewEnrichmentRegistry(),
-	}
-	l.Base = chain.NewBase(l, configs...)
-	return l
-}
-
-func (l *ARGEnrichmentLink) Params() []cfg.Param {
-	return []cfg.Param{}
-}
-
-func (l *ARGEnrichmentLink) Process(data outputters.NamedOutputData) error {
-	// Extract the Azure resource from the data
-	resource, ok := data.Data.(*model.AzureResource)
-	if !ok {
-		l.Logger.Debug("Skipping non-AzureResource data in enrichment", "data_type", fmt.Sprintf("%T", data.Data))
-		l.Send(data)
-		return nil
-	}
-
-	// Get template ID from resource properties
-	templateID, exists := resource.Properties["templateID"].(string)
-	if !exists {
-		l.Logger.Debug("No templateID found in resource properties, skipping enrichment", "resource_id", resource.Key)
-		l.Send(data)
-		return nil
-	}
-
-	// Enrich the resource with security testing commands
-	commands := l.registry.EnrichResource(l.Context(), templateID, resource)
-
-	if len(commands) > 0 {
-		l.Logger.Debug("Enriched resource with commands", "resource_id", resource.Key, "template_id", templateID, "command_count", len(commands))
-
-		// Add commands to resource properties
-		if resource.Properties == nil {
-			resource.Properties = make(map[string]any)
-		}
-		resource.Properties["commands"] = commands
-	}
-
-	// Send the enriched resource
-	l.Send(data)
-	return nil
 }
 
 // ARGTemplateLoaderLink loads and filters ARG templates by category
@@ -251,9 +50,9 @@ func (l *ARGTemplateLoaderLink) Process(input interface{}) error {
 	// - For modules with WithChainInputParam: string (subscription ID)
 	// - For modules with AzureSubscriptionGeneratorLink: string (subscription ID from generator)
 	l.Logger.Debug("ARGTemplateLoaderLink received input", "input", input, "type", fmt.Sprintf("%T", input))
-	
+
 	subscription := ""
-	
+
 	// If input is a string, it's a subscription ID from the chain
 	if inputStr, ok := input.(string); ok {
 		subscription = inputStr
@@ -262,13 +61,13 @@ func (l *ARGTemplateLoaderLink) Process(input interface{}) error {
 		// Fallback: get subscription from parameters for backward compatibility
 		subscriptions, err := cfg.As[[]string](l.Arg("subscription"))
 		l.Logger.Debug("subscription lookup", "subscriptions", subscriptions, "error", err, "all_args", l.Args())
-		
+
 		if len(subscriptions) > 0 {
 			subscription = subscriptions[0]
 		}
 		l.Logger.Debug("Using subscription from parameters", "subscription", subscription)
 	}
-	
+
 	l.Logger.Info("ARGTemplateLoaderLink starting", "subscription", subscription)
 
 	directory := ""
